chg: Refactored services
And also fixed some scoping

chg: Refactored services
