--- actix/src/services.rs
@@ -1,494 +1,0 @@
-// SPDX-FileCopyrightText: 2023-2026 Sayantan Santra <sayantan.santra689@gmail.com>
-// SPDX-License-Identifier: MIT
-
-use actix_files::NamedFile;
-use actix_session::Session;
-use actix_web::{
-    Either, HttpRequest, HttpResponse, Responder, delete, get,
-    http::StatusCode,
-    post, put,
-    web::{self, Redirect},
-};
-use argon2::{Argon2, PasswordVerifier, password_hash::PasswordHash};
-use log::{debug, info, warn};
-use serde::{Deserialize, Serialize};
-
-use crate::{
-    AppState,
-    auth::{self, is_session_valid},
-    config::{HashAlgorithm, SlugStyle},
-    database, utils,
-};
-
-// Error types
-pub enum ChhotoError {
-    ServerError,
-    ClientError { reason: String },
-}
-use ChhotoError::{ClientError, ServerError};
-
-// Define JSON struct for returning success/error data
-#[derive(Serialize)]
-pub struct JSONResponse {
-    pub success: bool,
-    pub error: bool,
-    pub reason: String,
-}
-
-// Define JSON struct for returning backend config
-#[derive(Serialize)]
-struct BackendConfig {
-    version: String,
-    site_url: Option<String>,
-    allow_capital_letters: bool,
-    public_mode: bool,
-    public_mode_expiry_delay: i64,
-    slug_style: String,
-    slug_length: usize,
-    try_longer_slug: bool,
-    frontend_page_size: u16,
-}
-
-// Needed to return the short URL to make it easier for programs leveraging the API
-#[derive(Serialize)]
-struct CreatedURL {
-    success: bool,
-    error: bool,
-    shorturl: String,
-    expiry_time: i64,
-}
-
-// Struct for returning information about a shortlink in expand
-#[derive(Serialize)]
-struct LinkInfo {
-    success: bool,
-    error: bool,
-    longurl: String,
-    hits: i64,
-    expiry_time: i64,
-    notes: String,
-}
-
-// Struct for query params in /api/all
-#[derive(Deserialize)]
-pub struct GetReqParams {
-    pub page_after: Option<String>,
-    pub page_no: Option<i64>,
-    pub page_size: Option<i64>,
-    pub filter: Option<String>,
-}
-
-// Define the routes
-
-// Add new links
-#[post("/api/new")]
-pub async fn add_link(
-    req: String,
-    data: web::Data<AppState>,
-    session: Session,
-    http: HttpRequest,
-) -> HttpResponse {
-    let config = &data.config;
-    // Call is_api_ok() function, pass HttpRequest
-    let result = auth::is_api_ok(http, config);
-    // If success, add new link
-    if result.success {
-        match utils::add_link(&req, &data.db, config, false) {
-            Ok((shorturl, expiry_time)) => {
-                let site_url = config.site_url.clone();
-                let shorturl = if let Some(url) = site_url {
-                    format!("{url}/{shorturl}")
-                } else {
-                    let protocol = if config.port == 443 { "https" } else { "http" };
-                    let port_text = if [80, 443].contains(&config.port) {
-                        String::new()
-                    } else {
-                        format!(":{}", config.port)
-                    };
-                    format!("{protocol}://localhost{port_text}/{shorturl}")
-                };
-                let response = CreatedURL {
-                    success: true,
-                    error: false,
-                    shorturl,
-                    expiry_time,
-                };
-                HttpResponse::Created().json(response)
-            }
-            Err(ServerError) => {
-                let response = JSONResponse {
-                    success: false,
-                    error: true,
-                    reason: "Something went wrong when adding the link.".to_string(),
-                };
-                HttpResponse::InternalServerError().json(response)
-            }
-            Err(ClientError { reason }) => {
-                let response = JSONResponse {
-                    success: false,
-                    error: true,
-                    reason,
-                };
-                HttpResponse::Conflict().json(response)
-            }
-        }
-    } else if result.error {
-        HttpResponse::Unauthorized().json(result)
-    // If password authentication or public mode is used - keeps backwards compatibility
-    } else {
-        let result = if auth::is_session_valid(session, config) {
-            utils::add_link(&req, &data.db, config, false)
-        } else if config.public_mode {
-            utils::add_link(&req, &data.db, config, true)
-        } else {
-            return HttpResponse::Unauthorized().body("Not logged in!");
-        };
-        match result {
-            Ok((shorturl, _)) => HttpResponse::Created().body(shorturl),
-            Err(ServerError) => HttpResponse::InternalServerError()
-                .body("Something went wrong when adding the link.".to_string()),
-            Err(ClientError { reason }) => HttpResponse::Conflict().body(reason),
-        }
-    }
-}
-
-// Return all active links
-#[get("/api/all")]
-pub async fn getall(
-    data: web::Data<AppState>,
-    session: Session,
-    params: web::Query<GetReqParams>,
-    http: HttpRequest,
-) -> HttpResponse {
-    let config = &data.config;
-    // Call is_api_ok() function, pass HttpRequest
-    let result = auth::is_api_ok(http, config);
-    // If success, return all links
-    if result.error {
-        HttpResponse::Unauthorized().json(result)
-    } else if result.success || auth::is_session_valid(session, config) {
-        match utils::getall(&data.db, params.into_inner()) {
-            Ok(s) => HttpResponse::Ok().body(s),
-            Err(ServerError) => HttpResponse::InternalServerError()
-                .body("Something went wrong while loading the links.".to_string()),
-            Err(ClientError { reason }) => HttpResponse::BadRequest().body(reason),
-        }
-    } else {
-        HttpResponse::Unauthorized().body("Not logged in!")
-    }
-}
-
-// Get information about a single shortlink
-#[post("/api/expand")]
-pub async fn expand(req: String, data: web::Data<AppState>, http: HttpRequest) -> HttpResponse {
-    let result = auth::is_api_ok(http, &data.config);
-    if result.success {
-        match database::find_url(&req, &data.db) {
-            Ok(chunks) => {
-                let body = LinkInfo {
-                    success: true,
-                    error: false,
-                    longurl: chunks.longlink,
-                    hits: chunks.hits,
-                    expiry_time: chunks.expiry_time,
-                    notes: chunks.notes,
-                };
-                HttpResponse::Ok().json(body)
-            }
-            Err(ServerError) => {
-                let body = JSONResponse {
-                    success: false,
-                    error: true,
-                    reason: "Something went wrong when finding the link.".to_string(),
-                };
-                HttpResponse::BadRequest().json(body)
-            }
-            Err(ClientError { reason }) => {
-                let body = JSONResponse {
-                    success: false,
-                    error: true,
-                    reason,
-                };
-                HttpResponse::BadRequest().json(body)
-            }
-        }
-    } else {
-        HttpResponse::Unauthorized().json(result)
-    }
-}
-
-// Get information about a single shortlink
-#[put("/api/edit")]
-pub async fn edit_link(
-    req: String,
-    session: Session,
-    data: web::Data<AppState>,
-    http: HttpRequest,
-) -> HttpResponse {
-    let config = &data.config;
-    let result = auth::is_api_ok(http, config);
-    if result.success || is_session_valid(session, config) {
-        match utils::edit_link(&req, &data.db, config) {
-            Ok(()) => {
-                let body = JSONResponse {
-                    success: true,
-                    error: false,
-                    reason: String::from("Edit was successful."),
-                };
-                HttpResponse::Created().json(body)
-            }
-            Err(ServerError) => {
-                let body = JSONResponse {
-                    success: false,
-                    error: true,
-                    reason: "Something went wrong when editing the link.".to_string(),
-                };
-                HttpResponse::InternalServerError().json(body)
-            }
-            Err(ClientError { reason }) => {
-                let body = JSONResponse {
-                    success: false,
-                    error: true,
-                    reason,
-                };
-                HttpResponse::BadRequest().json(body)
-            }
-        }
-    } else {
-        HttpResponse::Unauthorized().json(result)
-    }
-}
-
-// Get the site URL
-// This is deprecated, and might be removed in the future.
-// Use /api/getconfig instead
-#[get("/api/siteurl")]
-pub async fn siteurl(data: web::Data<AppState>) -> HttpResponse {
-    if let Some(url) = &data.config.site_url {
-        HttpResponse::Ok().body(url.clone())
-    } else {
-        HttpResponse::Ok().body("unset")
-    }
-}
-
-// Get the version number
-// This is deprecated, and might be removed in the future.
-// Use /api/getconfig instead
-#[get("/api/version")]
-pub async fn version() -> HttpResponse {
-    HttpResponse::Ok().body(format!("Chhoto URL v{}", utils::get_version()))
-}
-
-// Get the user's current role
-#[get("/api/whoami")]
-pub async fn whoami(
-    data: web::Data<AppState>,
-    session: Session,
-    http: HttpRequest,
-) -> HttpResponse {
-    let config = &data.config;
-    let result = auth::is_api_ok(http, config);
-    let acting_user = if result.success || is_session_valid(session, config) {
-        "admin"
-    } else if config.public_mode {
-        "public"
-    } else {
-        "nobody"
-    };
-    HttpResponse::Ok().body(acting_user)
-}
-
-// Get some useful backend config
-#[get("/api/getconfig")]
-pub async fn getconfig(
-    data: web::Data<AppState>,
-    session: Session,
-    http: HttpRequest,
-) -> HttpResponse {
-    let config = &data.config;
-    let result = auth::is_api_ok(http, config);
-    if result.success || is_session_valid(session, config) || data.config.public_mode {
-        let backend_config = BackendConfig {
-            version: utils::get_version(),
-            allow_capital_letters: config.allow_capital_letters,
-            public_mode: config.public_mode,
-            public_mode_expiry_delay: config.public_mode_expiry_delay.unwrap_or_default(),
-            site_url: config.site_url.clone(),
-            slug_style: (match config.slug_style {
-                SlugStyle::Uid => "UID",
-                SlugStyle::Pair => "Pair",
-            })
-            .to_string(),
-            slug_length: config.slug_length,
-            try_longer_slug: config.try_longer_slug,
-            frontend_page_size: config.frontend_page_size,
-        };
-        HttpResponse::Ok().json(backend_config)
-    } else {
-        HttpResponse::Unauthorized().json(result)
-    }
-}
-
-// 404 error page
-pub async fn error404() -> impl Responder {
-    NamedFile::open_async("./resources/static/404.html")
-        .await
-        .customize()
-        .with_status(StatusCode::NOT_FOUND)
-}
-
-// Handle a given shortlink
-#[get("/{shortlink}")]
-pub async fn link_handler(
-    shortlink: web::Path<String>,
-    data: web::Data<AppState>,
-) -> impl Responder {
-    let shortlink_str = shortlink.as_str();
-    if let Ok(longlink) = database::find_and_add_hit(shortlink_str, &data.db) {
-        if data.config.use_temp_redirect {
-            Either::Left(Redirect::to(longlink))
-        } else {
-            // Defaults to permanent redirection
-            Either::Left(Redirect::to(longlink).permanent())
-        }
-    } else {
-        Either::Right(
-            NamedFile::open_async("./resources/static/404.html")
-                .await
-                .customize()
-                .with_status(StatusCode::NOT_FOUND),
-        )
-    }
-}
-
-// Handle login
-#[post("/api/login")]
-pub async fn login(req: String, session: Session, data: web::Data<AppState>) -> HttpResponse {
-    let config = &data.config;
-    // Check if password is hashed using Argon2. More algorithms maybe added later.
-    let authorized = if let Some(password) = &config.password {
-        match config.hash_algorithm {
-            HashAlgorithm::Argon2 => {
-                debug!("Using Argon2 hash for password validation.");
-                let hash =
-                    PasswordHash::new(password).expect("The provided password hash is invalid.");
-                Some(
-                    Argon2::default()
-                        .verify_password(req.as_bytes(), &hash)
-                        .is_ok(),
-                )
-            }
-            HashAlgorithm::None => {
-                // If hashing is not enabled, use the plaintext password for matching
-                Some(password == &req)
-            }
-        }
-    } else {
-        None
-    };
-    if config.api_key.is_some() {
-        if let Some(valid_pass) = authorized
-            && !valid_pass
-        {
-            warn!("Failed login attempt!");
-            let response = JSONResponse {
-                success: false,
-                error: true,
-                reason: "Wrong password!".to_string(),
-            };
-            return HttpResponse::Unauthorized().json(response);
-        }
-        // Return Ok if no password was set on the server side
-        session
-            .insert("chhoto-url-auth", auth::gen_token_text())
-            .expect("Error inserting auth token.");
-
-        let response = JSONResponse {
-            success: true,
-            error: false,
-            reason: "Correct password!".to_string(),
-        };
-        info!("Successful login.");
-        HttpResponse::Ok().json(response)
-    } else {
-        // Keep this function backwards compatible
-        if let Some(valid_pass) = authorized
-            && !valid_pass
-        {
-            warn!("Failed login attempt!");
-            return HttpResponse::Unauthorized().body("Wrong password!");
-        }
-        // Return Ok if no password was set on the server side
-        session
-            .insert("chhoto-url-auth", auth::gen_token_text())
-            .expect("Error inserting auth token.");
-
-        info!("Successful login.");
-        HttpResponse::Ok().body("Correct password!")
-    }
-}
-
-// Handle logout
-// There's no reason to be calling this route with an API key
-#[delete("/api/logout")]
-pub async fn logout(session: Session) -> HttpResponse {
-    if session.remove("chhoto-url-auth").is_some() {
-        info!("Successful logout.");
-        HttpResponse::Ok().body("Logged out!")
-    } else {
-        HttpResponse::Unauthorized().body("You don't seem to be logged in.")
-    }
-}
-
-// Delete a given shortlink
-#[delete("/api/del/{shortlink}")]
-pub async fn delete_link(
-    shortlink: web::Path<String>,
-    data: web::Data<AppState>,
-    session: Session,
-    http: HttpRequest,
-) -> HttpResponse {
-    let config = &data.config;
-    // Call is_api_ok() function, pass HttpRequest
-    let result = auth::is_api_ok(http, config);
-    // If success, delete shortlink
-    if result.success {
-        match utils::delete_link(&shortlink, &data.db, data.config.allow_capital_letters) {
-            Ok(()) => {
-                let response = JSONResponse {
-                    success: true,
-                    error: false,
-                    reason: format!("Deleted {shortlink}"),
-                };
-                HttpResponse::Ok().json(response)
-            }
-            Err(ServerError) => {
-                let response = JSONResponse {
-                    success: false,
-                    error: true,
-                    reason: "Something went wrong when deleting the link.".to_string(),
-                };
-                HttpResponse::InternalServerError().json(response)
-            }
-            Err(ClientError { reason }) => {
-                let response = JSONResponse {
-                    success: false,
-                    error: true,
-                    reason,
-                };
-                HttpResponse::NotFound().json(response)
-            }
-        }
-    } else if result.error {
-        HttpResponse::Unauthorized().json(result)
-    // If using password - keeps backwards compatibility
-    } else if auth::is_session_valid(session, config) {
-        if utils::delete_link(&shortlink, &data.db, data.config.allow_capital_letters).is_ok() {
-            HttpResponse::Ok().body(format!("Deleted {shortlink}"))
-        } else {
-            HttpResponse::NotFound().body("Not found!")
-        }
-    } else {
-        HttpResponse::Unauthorized().body("Not logged in!")
-    }
-}
