+++ actix/src/services/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2023-2026 Sayantan Santra <sayantan.santra689@gmail.com>
+// SPDX-License-Identifier: MIT
+
+pub mod delete;
+pub mod get;
+pub mod post;
+pub mod put;
+pub mod types;
+pub mod utils;
